重构PGStorage,提取公共部分
