+++ pom.xml
@@ -21,6 +21,7 @@
         <module>example</module>
         <module>test</module>
         <module>assembly</module>
+        <module>dataSources/RelationAbstract</module>
     </modules>
 
     <distributionManagement>
